Cheaper paren_remover
